+++ test/test_CfgParser.hh
@@ -1,6 +1,6 @@
 //
 // test_CfgParser.hh for pekwm
-// Copyright (C) 2021-2022 Claes Nästén <pekdon@gmail.com>
+// Copyright (C) 2021-2023 Claes Nästén <pekdon@gmail.com>
 //
 // This program is licensed under the GNU GPL.
 // See the LICENSE file for more information.
@@ -31,7 +31,7 @@ public:
 
 TestCfgParser::TestCfgParser(void)
 	: TestSuite("CfgParser"),
-	  CfgParser()
+	  CfgParser("")
 {
 }