+++ test/test_Theme.hh
@@ -1,6 +1,6 @@
 //
 // test_Theme.cc for pekwm
-// Copyright (C) 2021-2022 Claes Nästén <pekdon@gmail.com>
+// Copyright (C) 2021-2023 Claes Nästén <pekdon@gmail.com>
 //
 // This program is licensed under the GNU GPL.
 // See the LICENSE file for more information.
@@ -73,7 +73,7 @@ TestTheme::loadColorMap(const char *cfg, std::map<int, int> &color_map)
 	CfgParserSourceString *source =
 		new CfgParserSourceString(":memory:", cfg);
 
-	CfgParser parser;
+	CfgParser parser("");
 	parser.parse(source);
 	CfgParser::Entry *section =
 		parser.getEntryRoot()->findSection("COLORMAPS");